services/api-gateway/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
PyJWT==2.10.1
bcrypt==4.2.1
email-validator==2.2.0
python-multipart==0.0.20
pytest==8.3.4
httpx==0.28.1
